Redo builder for leaf nodes in app example
