--- android/app-example/src/main/java/com/badoo/ribs/example/rib/blocker/builder/BlockerScope.kt
@@ -1,7 +1,0 @@
-package com.badoo.ribs.example.rib.blocker.builder
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-internal annotation class BlockerScope
