--- android/app-example/src/main/java/com/badoo/ribs/example/rib/lorem_ipsum/builder/LoremIpsumScope.kt
@@ -1,7 +1,0 @@
-package com.badoo.ribs.example.rib.lorem_ipsum.builder
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-internal annotation class LoremIpsumScope
