+++ android/app-example/src/main/java/com/badoo/ribs/example/rib/menu/MenuBuilder.kt
@@ -1,0 +1,38 @@
+package com.badoo.ribs.example.rib.menu
+
+import android.os.Bundle
+import com.badoo.ribs.core.Builder
+import com.badoo.ribs.core.Node
+import com.badoo.ribs.customisation.customisationsBranchFor
+import com.badoo.ribs.customisation.getOrDefault
+import com.badoo.ribs.example.rib.menu.feature.MenuFeature
+
+class MenuBuilder(
+    dependency: Menu.Dependency
+) : Builder<Menu.Dependency>() {
+
+    override val dependency : Menu.Dependency = object : Menu.Dependency by dependency {
+        override fun ribCustomisation() = dependency.customisationsBranchFor(Menu::class)
+    }
+
+    fun build(savedInstanceState: Bundle?): Node<MenuView> {
+        val customisation = dependency.getOrDefault(Menu.Customisation())
+        val router = MenuRouter(savedInstanceState)
+        val feature = MenuFeature()
+        val interactor = MenuInteractor(
+            savedInstanceState,
+            router,
+            dependency.menuInput(),
+            dependency.menuOutput(),
+            feature
+        )
+
+        return Node(
+            savedInstanceState = savedInstanceState,
+            identifier = object : Menu {},
+            viewFactory = customisation.viewFactory(null),
+            router = router,
+            interactor = interactor
+        )
+    }
+}
