--- android/app-example/src/main/java/com/badoo/ribs/example/rib/menu/builder/MenuScope.kt
@@ -1,7 +1,0 @@
-package com.badoo.ribs.example.rib.menu.builder
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-internal annotation class MenuScope
