--- android/app-example/src/main/java/com/badoo/ribs/example/rib/portal_overlay_test/builder/PortalOverlayTestScope.kt
@@ -1,7 +1,0 @@
-package com.badoo.ribs.example.rib.portal_overlay_test.builder
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-internal annotation class PortalOverlayTestScope
